+++ examples/run_img_concat.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/sh
 cd ..
-python nmag/image/img_concat.py --input_folder examples/img_concat --out_path examples/img_concat.png
-python nmag/image/img_concat.py --input_folder examples/img_concat --out_path examples/img_concat.pdf
+python nmag/image/img_concat.py --src_dir examples/img_concat --dst_path examples/img_concat.png
+python nmag/image/img_concat.py --src_dir examples/img_concat --dst_path examples/img_concat.pdf