+++ examples/run_md_rename.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 cd ..
-python nmag/mmage/md_renamer.py --file_path "examples/md_rename/DimensionReduction.md" --new_name test.md --output_folder examples/md_rename
+python nmag/mmage/md_renamer.py --src_path "examples/mds/DimensionReduction.md" --dst_path examples/mds/o.md