+++ examples/run_pdfext.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 cd ..
-python nmag/pmage/pdfext.py --in_path examples/pdfs/OR2_03_BnBandHeuristic.pdf --out_path examples/pdfs/OR2_03_BnBandHeuristic_.pdf --range "[0, 15]"
+python nmag/pmage/pdfext.py --src_path examples/pdfs/OR2_03_BnBandHeuristic.pdf --dst_path examples/pdfs/OR2_03_BnBandHeuristic_.pdf --range "[0, 15]"