+++ nmag/pmage/pdf2img.py
@@ -8,14 +8,14 @@ sys.path.append(osp.abspath(osp.join(osp.dirname(__file__), "../")))
 from utils import *
 
 
-def pdf_to_img(input_path, output_folder):
-    images = convert_from_path(input_path)
+def pdf_to_img(src_path, dst_dir):
+    images = convert_from_path(src_path)
 
-    if not osp.exists(output_folder):
-        os.makedirs(output_folder)
+    if not osp.exists(dst_dir):
+        os.makedirs(dst_dir)
 
     for i in range(len(images)):
-        path = osp.join(output_folder, "p" + str(i) + ".jpg")
+        path = osp.join(dst_dir, "p" + str(i) + ".jpg")
         images[i].save(path, "JPEG")
 
 
@@ -23,15 +23,15 @@ if __name__ == "__main__":
     logger = get_logger(filename="slide2md.log", verb_level="info", method="w2file")
     parser = argparse.ArgumentParser()
     parser.add_argument(
-        "--input_path",
+        "--src_path",
         type=str,
         default="examples/pdf2img/OR2_03_BnBandHeuristic.pdf",
         help="the folder that contains images",
     )
     parser.add_argument(
-        "--output_folder",
+        "--dst_dir",
         type=str,
         default="examples/pdf2img/OR2_03_BnBandHeuristic",
     )
     args = parser.parse_args()
-    pdf_to_img(args.input_path, args.output_folder)
+    pdf_to_img(args.src_path, args.dst_dir)