+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 ### Feat
 
 - add setup.py and commitizen for changelog automation and move img_concat to image folder, add a skeleton for cli
-- add output_folder input arg in renamer
+- add dst_dir input arg in renamer
 - add args parser in both typora functions